backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.1
alembic==1.13.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
pydantic==1.10.13
telnyx==2.1.1
openai==1.54.0
httpx==0.25.1
stripe>=10,<13
aiofiles==23.2.1
email-validator==2.1.0
pytz==2024.1
